nulled

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: babonnet <babonnet@42angouleme.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/30 19:33:17 by bbonnet           #+#    #+#             */
-/*   Updated: 2023/11/06 13:12:46 by babonnet         ###   ########.fr       */
+/*   Updated: 2023/11/09 19:17:27 by babonnet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	const char	*start;
